0.19.8 new package for Inventory + refactoring
