+++ wfrp/src/main/java/com/warhammer/wfrp/inventory/InventoryController.java
@@ -1,8 +1,6 @@
-package com.warhammer.wfrp.controller;
+package com.warhammer.wfrp.inventory;
 
 import com.warhammer.wfrp.dto.inventory.HirelingDto;
-import com.warhammer.wfrp.dto.inventory.InventoryDto;
-import com.warhammer.wfrp.service.InventoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
