+++ wfrp/src/main/java/com/warhammer/wfrp/inventory/InventoryRepository.java
@@ -1,6 +1,5 @@
-package com.warhammer.wfrp.repository.inventory;
+package com.warhammer.wfrp.inventory;
 
-import com.warhammer.wfrp.model.inventory.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface InventoryRepository extends JpaRepository<Inventory, Integer> {
